1.登录时服务器返回角色权限,token附带权限 2.用户删除时验证权限 3.前端添加操作权限

+++ test02/nodeServer/routes/users.js
@@ -4,12 +4,18 @@ var userService = require('../services/userService')
 
 const loggerFactory = require('../log/log4js')
 const logger = loggerFactory.getLogger('login')
-  /* GET users listing. */
-router.get('/getAllUsers', async function(req, res, next) {
+/* GET users listing. */
+router.get('/getAllUsers', async function (req, res, next) {
   console.log('响应/getAllUsers');
   console.log(req.user);
-  
-  const result = await userService.getAllUsers()
+  let result
+  if (req.query.page && req.query.size) {
+    console.log(`page and size ${req.query.page} ${req.query.size}`);
+    result = await userService.getUsersByPage({ page: req.query.page, size: req.query.size })
+  } else {
+    result = await userService.getAllUsers()
+  }
+
   if (!result.err) {
     res.send(result)
   } else {
@@ -18,7 +24,7 @@ router.get('/getAllUsers', async function(req, res, next) {
   }
 });
 
-router.post('/register', async function(req, res, next) {
+router.post('/register', async function (req, res, next) {
   let username = req.body.username
   let password = req.body.password
   let mail = req.body.mail
@@ -31,7 +37,7 @@ router.post('/register', async function(req, res, next) {
   res.send(msg)
 });
 
-router.post('/update', async function(req, res, next) {
+router.post('/update', async function (req, res, next) {
   let user = req.body
 
   console.log(`${JSON.stringify(req.body)} 响应/update`);
@@ -42,14 +48,17 @@ router.post('/update', async function(req, res, next) {
   res.send(msg)
 });
 
-router.post('/delete', async function(req, res, next) {
-  let id = req.body
-
+router.post('/delete', async function (req, res, next) {
   console.log(`${JSON.stringify(req.body)} 响应/delete`);
+  let id = req.body
+  if (req.user.remark !== '管理员') {
+    console.log('没有权限,非法操作');
+    res.json({ message: '没有权限', errcode: 4 })
+  } else {
+    const msg = await userService.deleteUserById(id)
+    console.log(msg);
 
-  const msg = await userService.deleteUserById(id)
-  console.log(msg);
-
-  res.send(msg)
+    res.send(msg)
+  }
 })
 module.exports = router;
